+++ mobile/src/app/rm-tabs/clients/rm-clients.module.ts
@@ -5,11 +5,12 @@ import { FormsModule } from '@angular/forms';
 import { IonicModule } from '@ionic/angular';
 import { RmClientsPage } from './rm-clients.page';
 import { RmClientEditSheetModule } from '../../features/rm/client-edit/rm-client-edit-sheet.module';
+import { RmCollectorAssignSheetModule } from '../../features/rm/collector-assign/rm-collector-assign-sheet.module';
 
 const routes: Routes = [{ path: '', component: RmClientsPage }];
 
 @NgModule({
-  imports: [CommonModule, FormsModule, IonicModule, RouterModule.forChild(routes), RmClientEditSheetModule],
+  imports: [CommonModule, FormsModule, IonicModule, RouterModule.forChild(routes), RmClientEditSheetModule, RmCollectorAssignSheetModule],
   declarations: [RmClientsPage]
 })
 export class RmClientsPageModule {}